[NRRPLT-4577] Camera does not publish after world reset.
The iCub visual tracking experiment would fail after a reset because
the gazebo_ros_camera plugin was not properly updating/using the
last publication time after a reset. If you ran a simulation for 15
seconds and then reset, the camera would not publish again until after
15 seconds of new sim time had elapsed (e.g. the clock had to go back
up to 15 seconds). This was sufficient time for the iCub's eyes to
move so far that it couldn't see the screen and respond to stimulus
anymore. This issue has likely been present for the lifetime of the
platform.

This patch resolves the issue by using the parent sensor object which
is properly updated within the Gazebo code. I will submit a pull
request / bug report as this issue is still present in the latest
upstream gazebo_ros_packages.

Change-Id: I6756971580fc94fab272692dbe7506f756fe9e00