[NRRPLT-4676] Add service to wait for rendering scene to be loaded.
This patch adds a new service call to wait for the rendering scene
to match the world (loaded models/lights). This is required to ensure
that simulations that need a rendering scene (e.g. those with cameras or
other environment sensors and not purely math based models) are ready
before the user can start a simulation. This can take a few seconds,
especially on the vizcluster where a node will be at 100% cpu usage for
a few seconds before finally loading all models and textures into the
rendering scene and then finally being able to publish camera images.

This should only really take more than a second or two on the initial
simulation load, but [NRRPLT-4699] has been created to investigate the
need to call this after each object insertion.

This patch also includes minor changes to the icub modules to defer
potentially blocking calls from Init to Load, though this shouldn't
change much performance-wise.

Change-Id: I4d7b0a6d81272822dc5a6944c4eae51e1b33b706