[NRRPLT-4663] Scale field in ModelState message can't be zero.
Zero valued scaling factors in ModelState messages, now, makes the call terminate
unsuccesfully with a status message.

A zero scaling factor can be specified by the user in TFs or, implicitely,
in case the field is left unspecified, being it the default value.

ROS messages don't allow to specify default values for user-defined types
(Scale is a geometry_msgs/Vector3).

Change-Id: I77dcdcd71cff4d411b913f98f531c996f89cc0ef