[NRRPLT-3910] Plugin for synchronized ROS/Gazebo recording.
This commit adds a new plugin to support dynamically recording both
Gazebo and ROS recordings via ROS service calls.

The following service calls are provided:
- /gazebo/recording/start - Start coordinated recording.
- /gazebo/recording/stop - Stop recording and wait for files to be
                           completely written.
- /gazebo/recording/cancel - Cancel current recording and discard files.
- /gazebo/recording/cleanup - Clean all generated files and recordings
                              to be used for reset/shutdown.
- /gazebo/recording/get_recording - Notifies the caller if recording is
                                    started and the output path.

Recordings can be generated on demand with multiple calls to start and
stop that will generate several files. These files are saved in a temp
dir that can be retrieved by the caller in the form:
- <dir>/gzserver/<num>.log - Gazebo state logs
- <dir>/ros/<num>.bag - ROS bags

The plugin also issues cleanup commands on simulation time resets and
ensures the recordings are synchronized by properly managing simulation
state when starting/stopping recording.

Change-Id: I5a82f8ec92bc15c63aeb9fd0a214dcbf87457691